+++ docs/release_notes.rst
@@ -2,6 +2,11 @@
 .. role:: smaller
 .. role:: noteversion
 
+Version 0.3.5 :small:`13 June 2019`
+---------------------------------------
+Fixed a bug that was prone to produce infinities during reconstruction.
+Improved plotting functions and added new ones for plotting mapped cells.
+
 Version 0.3.4 :small:`27 February 2019`
 ---------------------------------------
 novoSpaRc reconstructs single-cell gene expression without relying on existing